refactor integrated test
