+++ shenyu-integrated-test/pom.xml
@@ -31,11 +31,11 @@
 
     <groupId>org.apache.shenyu</groupId>
     <artifactId>shenyu-integrated-test</artifactId>
-    <version>2.3.1-SNAPSHOT</version>
+    <version>2.4.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <properties>
-        <shenyu.version>2.3.1-SNAPSHOT</shenyu.version>
+        <shenyu.version>2.4.0-SNAPSHOT</shenyu.version>
         <maven-compiler-plugin.version>3.7.0</maven-compiler-plugin.version>
         <versions-maven-plugin.version>2.5</versions-maven-plugin.version>
     </properties>
