+++ shenyu-integrated-test/shenyu-integrated-test-http/docker-compose.yml
@@ -69,15 +69,15 @@ services:
       - shenyu
 
   shenyu-admin:
-    image: dromara/shenyu-admin:2.3.1-SNAPSHOT
+    image: dromara/shenyu-admin:2.4.0-SNAPSHOT
     container_name: shenyu-admin
     restart: always
     networks:
       - shenyu
-#    volumes:
-#      - type: bind
-#        source: ./admin-config/application.yml
-#        target: /opt/shenyu-admin/conf/application.yml
+    volumes:
+      - type: bind
+        source: ./admin-config/schema.sql
+        target: /opt/shenyu-admin/conf/schema.sql
     depends_on:
       shenyu-mysql:
         condition: service_healthy
@@ -89,15 +89,16 @@ services:
       - SPRING_PROFILES_ACTIVE=mysql
       - "spring.datasource.url=jdbc:mysql://shenyu-mysql:3306/shenyu?useUnicode=true&characterEncoding=utf-8&useSSL=false"
       - "spring.datasource.password=root"
-      - "soul.sync.http.url=http://shenyu-admin:9095/http"
+      - "shenyu.sync.http.url=http://shenyu-admin:9095/http"
+      - "shenyu.database.custom_init_script=/opt/shenyu-admin/conf/schema.sql"
     healthcheck:
       test: [ "CMD", "wget", "http://shenyu-admin:9095/actuator/health" ]
       timeout: 2s
       retries: 30
 #  from http plugin dockerfile
-  shenyu-bootstrap:
-    container_name: shenyu-bootstrap
-    image: dromara/shenyu-bootstrap:2.3.1-SNAPSHOT
+  shenyu-integrated-test-http:
+    container_name: shenyu-integrated-test-http
+    image: dromara/shenyu-integrated-test-http:latest
     restart: always
     deploy:
       resources:
@@ -106,7 +107,7 @@ services:
 #    volumes:
 #      - type: bind
 #        source: ./bootstrap-config
-#        target: /opt/shenyu-bootstrap/conf/
+#        target: /opt/shenyu-integrated-test-http/conf/
     depends_on:
       shenyu-examples-http:
         condition: service_healthy
