+++ shenyu-integrated-test/shenyu-integrated-test-http/src/main/java/org/apache/shenyu/integratedtest/App.java
@@ -17,11 +17,15 @@
 
 package org.apache.shenyu.integratedtest;
 
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
 /**
  * Hello world!
  */
+@SpringBootApplication
 public class App {
     public static void main(String[] args) {
-        System.out.println("Hello World!");
+        SpringApplication.run(App.class);
     }
 }
