+++ shenyu-integrated-test/shenyu-integrated-test-http/src/test/java/org/apache/shenyu/integratedtest/http/dto/AdminResponse.java
@@ -14,25 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.shenyu.integratedtest.http.dto;
 
-package org.apache.shenyu.integratedtest.dto;
+import lombok.Getter;
+import lombok.Setter;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-/**
- * The type Order dto.
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderDTO implements Serializable {
-
-    private String id;
-
-    private String name;
-
+@Getter
+@Setter
+public class AdminResponse<T> {
+    private Integer code;
+    private String message;
+    private T data;
 }
