+++ shenyu-integrated-test/shenyu-integrated-test-http/src/test/java/org/apache/shenyu/integratedtest/http/dto/UserDTO.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.integratedtest.dto;
+package org.apache.shenyu.integratedtest.http.dto;
 
 import lombok.Data;
